requirements.txt
mcp>=1.0.0
schwabdev>=3.0.0
pydantic>=2.0
diskcache>=5.6
python-dotenv>=1.0
httpx>=0.24
